+++ CHANGES.md
@@ -6,6 +6,7 @@
 - Move map credits to map column so it don't get hidden by chart panel
 - TSify `MapColumn` module and reorganize components directory structure.
 - Add null check to `WebMapServiceCatalogItem` `rectangle` calculation - and now we ascend tree of WMS `Layers` until we find a rectangle.
+- Add `includeMembersRegex` to `GroupTraits`. This can be used to filter group members by id/name using a regular expression.
 - [The next improvement]
 
 #### next release (8.3.3)